+++ sample/main.cpp
@@ -8,42 +8,42 @@
 #include "ZFontIcon/ZFont_fa6.h"
 #include "ZFontIcon/ZFont_fa5.h"
 #include "ZFontIcon/ZFont_fa4.h"
+#include "ZFontIcon/ZFont_bi1.h"
 #include "ZFontIcon/ZFont_gmi4.h"
+#include "ZFontIcon/ZFont_mdi6.h"
+#include "ZFontIcon/ZFont_ph1.h"
 
-// Enable Font Awesome 6 Preview and Font Awesome 5 Pro
-bool USE_FA6_FREE_PREVIEW= false; // Note: Edit resource file in plain text Editor to add FA6 Free fonts
-bool USE_FA6_PRO_PREVIEW=  false; // Note: Edit resource file in plain text Editor to add FA6 Pro fonts
-bool USE_FA5_PRO_FONT=     false; // Note: Edit resource file in plain text Editor to add FA5 Pro fonts
+// Enable Font Awesome 6 Pro and Font Awesome 5 Pro
+bool USE_FA6_PRO_FONT= false; // Note: Edit resource file in plain text Editor to add FA6 Pro fonts
+bool USE_FA5_PRO_FONT= false; // Note: Edit resource file in plain text Editor to add FA5 Pro fonts
 
 
-//------------------------------
-// Fonts Awesome 6.x (Preview) -
-//------------------------------
-void registerFA6_fonts() {
+//--------------------
+// Fonts Awesome 6.x -
+//--------------------
+void registerFa6_fonts() {
     // Note that:
     //   Solid, Regular Light and Thin styles share the same family name,
     //   Brands style have a different family name
 
-    if(USE_FA6_FREE_PREVIEW) {
-        // FA6 Free (Solid and Regular styles)
-        ZFontIcon::addFont(":/fa6/" + Fa6::FA6_OTF_FILE_FREE_SOLID);
-        ZFontIcon::addFont(":/fa6/" + Fa6::FA6_OTF_FILE_FREE_REGULAR); // Support less icons than Solid style (~150)
-        // FA6 Brands (Regular style)
-        ZFontIcon::addFont(":/fa6/" + Fa6brands::FA6_OTF_FILE_BRANDS);
-    }
+    // FA6 Free (Solid and Regular styles)
+    ZFontIcon::addFont(":/fa6/" + Fa6::FA6_TTF_FILE_FREE_SOLID);
+    ZFontIcon::addFont(":/fa6/" + Fa6::FA6_TTF_FILE_FREE_REGULAR); // Support less icons than Solid style (~150)
+    // FA6 Brands (Regular style)
+    ZFontIcon::addFont(":/fa6/" + Fa6brands::FA6_TTF_FILE_BRANDS);
 
     // FA6 Pro (Solid, Regular Light and Thin styles)
-    if(USE_FA6_PRO_PREVIEW) {
-        ZFontIcon::addFont(":/fa6/" + Fa6pro::FA6_OTF_FILE_PRO_SOLID);
-        ZFontIcon::addFont(":/fa6/" + Fa6pro::FA6_OTF_FILE_PRO_REGULAR);
-        ZFontIcon::addFont(":/fa6/" + Fa6pro::FA6_OTF_FILE_PRO_LIGHT);
-        ZFontIcon::addFont(":/fa6/" + Fa6pro::FA6_OTF_FILE_PRO_THIN);
+    if(USE_FA6_PRO_FONT) {
+        ZFontIcon::addFont(":/fa6/" + Fa6pro::FA6_TTF_FILE_PRO_SOLID);
+        ZFontIcon::addFont(":/fa6/" + Fa6pro::FA6_TTF_FILE_PRO_REGULAR);
+        ZFontIcon::addFont(":/fa6/" + Fa6pro::FA6_TTF_FILE_PRO_LIGHT);
+        ZFontIcon::addFont(":/fa6/" + Fa6pro::FA6_TTF_FILE_PRO_THIN);
     }
 }
 
 
 //-----------------------
-// Fonts Awesome 5.15.3 -
+// Fonts Awesome 5.15.4 -
 //-----------------------
 void registerFa5_fonts() {
     // Note that:
@@ -51,16 +51,16 @@ void registerFa5_fonts() {
     //   Brands style have a different family name
 
     // FA5 Free (Solid and Regular styles)
-    ZFontIcon::addFont(":/fa5/" + Fa5::FA5_OTF_FILE_FREE_SOLID);
-    ZFontIcon::addFont(":/fa5/" + Fa5::FA5_OTF_FILE_FREE_REGULAR); // Support less icons than Solid style (~150)
+    ZFontIcon::addFont(":/fa5/" + Fa5::FA5_WOFF_FILE_FREE_SOLID);
+    ZFontIcon::addFont(":/fa5/" + Fa5::FA5_WOFF_FILE_FREE_REGULAR); // Support less icons than Solid style (~150)
     // FA5 Brands (Regular style)
-    ZFontIcon::addFont(":/fa5/" + Fa5brands::FA5_OTF_FILE_BRANDS);
+    ZFontIcon::addFont(":/fa5/" + Fa5brands::FA5_WOFF_FILE_BRANDS);
 
     // FA5 Pro (Solid, Regular Light and Thin styles)
     if(USE_FA5_PRO_FONT) {
-        ZFontIcon::addFont(":/fa5/" + Fa5pro::FA5_OTF_FILE_PRO_SOLID);
-        ZFontIcon::addFont(":/fa5/" + Fa5pro::FA5_OTF_FILE_PRO_REGULAR);
-        ZFontIcon::addFont(":/fa5/" + Fa5pro::FA5_OTF_FILE_PRO_LIGHT);
+        ZFontIcon::addFont(":/fa5/" + Fa5pro::FA5_WOFF_FILE_PRO_SOLID);
+        ZFontIcon::addFont(":/fa5/" + Fa5pro::FA5_WOFF_FILE_PRO_REGULAR);
+        ZFontIcon::addFont(":/fa5/" + Fa5pro::FA5_WOFF_FILE_PRO_LIGHT);
     }
 }
 
@@ -69,7 +69,15 @@ void registerFa5_fonts() {
 // Fonts Awesome 4.7 (legacy) -
 //-----------------------------
 void registerFa4_fonts() {
-    ZFontIcon::addFont(":/fa4/"  + Fa4::FA4_OTF_FILE_FREE_REGULAR);
+    ZFontIcon::addFont(":/fa4/"  + Fa4::FA4_WOFF_FILE_REGULAR);
+}
+
+
+//----------------------
+// BootStrap Icons 1.x -
+//----------------------
+void registerBi1_fonts() {
+    ZFontIcon::addFont(":/bi1/"  + Bi1::BI1_WOFF_FILE_REGULAR);
 }
 
 
@@ -83,6 +91,22 @@ void registerGmi4_fonts() {
 }
 
 
+//----------------------------
+// Material Design Icons 6.x -
+//----------------------------
+void registerMdi6_fonts() {
+    ZFontIcon::addFont(":/mdi6/" + Mdi6::MDI6_WOFF_FILE_REGULAR);
+}
+
+
+//----------------------------
+// Phosphor 1.x -
+//----------------------------
+void registerPh1_fonts() {
+    ZFontIcon::addFont(":/ph1/" + Ph1::PH1_WOFF_FILE_REGULAR);
+}
+
+
 
 
 
@@ -90,10 +114,13 @@ int main(int argc, char *argv[]) {
     QApplication app(argc, argv);
 
     // Register fonts
-    registerFA6_fonts();
+    registerFa6_fonts();
     registerFa5_fonts();
     registerFa4_fonts();
+    registerBi1_fonts();
     registerGmi4_fonts();
+    registerMdi6_fonts();
+    registerPh1_fonts();
 
     MainWindow mainWindow;
     mainWindow.show();